# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zellerfeld Assignment - Twitter Clone API",
        "description": "This API provides endpoints for a Twitter clone application, allowing users to manage posts and authenticate.",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in a user",
                "description": "Logs in a user and generates an access token.",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged in successfully"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out a user",
                "description": "Logs out the currently authenticated user by revoking their token.",
                "operationId": "085014c580531cf4f0a1c53b7c3faa65",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogoutRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Retrieve the authenticated user's profile",
                "description": "Retrieves the profile information of the currently authenticated user.",
                "operationId": "3c91ae2e201145a495a0b5696c6b8901",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of the user's profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "The user's ID",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "The user's name",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "The user's email address",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the authenticated user's profile",
                "description": "Updates the profile information of the currently authenticated user.",
                "operationId": "354f5b23296bed4036564563e7423f43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "The new email address",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "name": {
                                        "description": "The new name",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "password": {
                                        "description": "The new password",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful update of the user's profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "The user's ID",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "The updated name",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "The updated email address",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The user's email address",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The user's password",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "LogoutRequest": {
                "description": "Schema for logout request",
                "type": "object"
            },
...
        }
    },
...
}